Use autoconf ability to compile in a builddir. #include "config.h" has been replaced by #include <config.h> so that it will be searched. Added -I. to CFLAGS, to search for <config.h>. Adapted autocon/acinclude.m4 CS_CHECK_VERSION to search in ac_dir instead of builddir for svn/git/hg revision
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@3669 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ autoconf/acinclude.m4
@@ -279,21 +279,23 @@ fi
 ## -----------------------------##
 AC_DEFUN([PATCH_LIBTOOL_TO_ADD_HOST_CC],
 [# patch libtool to add HOST_CC sometimes needed in crosscompiling a win32 dll
-if grep "HOST_CC" libtool >/dev/null; then
-  if test "$build" != "$host" ; then
-    if test "_$HOST_CC" = "_" ; then
-      HOST_CC="false"
-      for i in `echo $PATH | sed 's,:, ,g'` ; do
-      test -x $i/cc && HOST_CC=$i/cc
-      done
+if test -f libtool; then
+  if grep "HOST_CC" libtool >/dev/null; then
+    if test "$build" != "$host" ; then
+      if test "_$HOST_CC" = "_" ; then
+        HOST_CC="false"
+        for i in `echo $PATH | sed 's,:, ,g'` ; do
+        test -x $i/cc && HOST_CC=$i/cc
+        done
+      fi
+  AC_MSG_RESULT(patching libtool to add HOST_CC=$HOST_CC)
+      test -f libtool.old || (mv libtool libtool.old && cp libtool.old libtool)
+      sed -e "/BEGIN.*LIBTOOL.*CONFIG/a\\
+  HOST_CC=$HOST_CC" libtool >libtool.new 
+      (test -s libtool.new || rm libtool.new) 2>/dev/null
+      test -f libtool.new && mv libtool.new libtool # not 2>/dev/null !!
+      test -f libtool     || mv libtool.old libtool
     fi
-AC_MSG_RESULT(patching libtool to add HOST_CC=$HOST_CC)
-    test -f libtool.old || (mv libtool libtool.old && cp libtool.old libtool)
-    sed -e "/BEGIN.*LIBTOOL.*CONFIG/a\\
-HOST_CC=$HOST_CC" libtool >libtool.new 
-    (test -s libtool.new || rm libtool.new) 2>/dev/null
-    test -f libtool.new && mv libtool.new libtool # not 2>/dev/null !!
-    test -f libtool     || mv libtool.old libtool
   fi
 fi  
 ])  
@@ -372,40 +374,43 @@ AC_DEFUN([CS_GET_VERSION], [
   SCCP_VERSION="unknown"
   SCCP_REVISION="unknown"
 
-  if test -d .svn || test -d ../.svn || test -d ../../.svn;then
-    if test "`basename $(svn info .| grep -i ^URL: | cut -d ' ' -f 2)`" != "trunk"; then
-        SCCP_VERSION=`basename $(svnpath)`
-        SCCP_BRANCH=`basename $(svnpath)`
-    else
-      if test -f .version; then
-        SCCP_VERSION="`cat .version|cut -d_ -f1`"
-        SCCP_BRANCH="`cat .version|cut -d_ -f2`"
-      else 
-        SCCP_VERSION="TRUNK"
-        SCCP_BRANCH="TRUNK"
+  BASE=`dirname $ac_dir`
+  pushd $BASE
+  if test ! -z "$SVN" && test ! -z "$SVNVERSION" && (test -d .svn || test -d ../.svn || test -d ../../.svn);then
+      if test "`basename $($SVN info .| grep -i ^URL: | cut -d ' ' -f 2)`" != "trunk"; then
+          SCCP_VERSION=`basename $(svnpath)`
+          SCCP_BRANCH=`basename $(svnpath)`
+      else
+        if test -f .version; then
+          SCCP_VERSION="`cat .version|cut -d_ -f1`"
+          SCCP_BRANCH="`cat .version|cut -d_ -f2`"
+        else 
+          SCCP_VERSION="TRUNK"
+          SCCP_BRANCH="TRUNK"
+        fi
       fi
-    fi
-    SCCP_REVISION="`svnversion . |cut -dM -f1`"
-  elif test -d .hg || test -d ../.hg || test -d ../../.hg;then
-    SCCP_VERSION="`cat .hg/hgrc |grep http|awk '{print $3}'|tr "/" "\n"|tail -n2|head -n1`"
-    if grep -q "default" .hg/branch; then
-      SCCP_BRANCH="TRUNK"
-    else 
-      SCCP_BRANCH="`cat .hg/branch`"
-    fi
-    SCCP_REVISION="`hg parents|grep changeset|tail -n1|sed 's/\(.*:\)\(.*\)\(:.*\)/\2/g'|sed 's/\ //g'`"
-  elif test -d .git || test -d ../.git || test -d ../../.git;then
-    SCCP_VERSION="`cat .version|cut -d_ -f1`"
-dnl    if test "`git branch | awk '{print $2}' | sed 's/\n//g'`" = "master"; then
-      SCCP_BRANCH="TRUNK"
-dnl    else 
-dnl      SCCP_BRANCH="`git branch | awk '{print $2}' | sed s/\n//g`"
-dnl    fi
-    SCCP_REVISION="`git describe --always`"
+      SCCP_REVISION="`svnversion .|cut -d: -f2 |cut -dM -f1`"
+  elif test ! -z "$HG" && (test -d .hg || test -d ../.hg || test -d ../../.hg);then
+      SCCP_VERSION="`cat .hg/hgrc |grep http|awk '{print $3}'|tr "/" "\n"|tail -n2|head -n1`"
+      if grep -q "default" .hg/branch; then
+        SCCP_BRANCH="TRUNK"
+      else 
+        SCCP_BRANCH="`cat .hg/branch`"
+      fi
+      SCCP_REVISION="`hg parents|grep changeset|tail -n1|sed 's/\(.*:\)\(.*\)\(:.*\)/\2/g'|sed 's/\ //g'`"
+  elif test ! -z "$GIT" && (test -d .git || test -d ../.git || test -d ../../.git);then
+      SCCP_VERSION="`cat .version|cut -d_ -f1`"
+  dnl    if test "`git branch | awk '{print $2}' | sed 's/\n//g'`" = "master"; then
+        SCCP_BRANCH="TRUNK"
+  dnl    else 
+  dnl      SCCP_BRANCH="`git branch | awk '{print $2}' | sed s/\n//g`"
+  dnl    fi
+      SCCP_REVISION="`git describe --always`"
   elif test -f .version; then
     SCCP_BRANCH="`cat .version|cut -d_ -f2`"
     SCCP_VERSION="`cat .version|cut -d_ -f1`"
   fi
+  popd
 
   AC_DEFINE_UNQUOTED([SCCP_VERSION],  "${SCCP_VERSION}", [Define the SCCP Version])
   AC_DEFINE_UNQUOTED([SCCP_BRANCH],   "${SCCP_BRANCH}", [Define the SCCP Branch])
